New Feature: Portfolio Project Enhancements
Added a Projects screen with a rotating 3D box animation to showcase projects.
Implemented a flip card for the service section on the home screen, providing detailed information on both sides.
Enhanced the contact page with a flip card animation and structured skill information.
Improved navigation by adding a direct link to the Projects screen in the drawer and floating action button.
Organized skill information with icons and detailed descriptions in both the service and contact sections.